BJS2-34350
refactoring code
