+++ src/main/java/school/faangSprint2/t13/Main.java
@@ -1,0 +1,27 @@
+package school.faangSprint2.t13;
+
+import java.util.function.BiFunction;
+
+public class Main {
+    public static void main(String[] args) {
+        FareCalculator fareCalculator = new FareCalculator();
+
+// Определение тарифов
+        BiFunction<Double, Double, Double> economyFare = (distance, time) -> 1.0 * distance + 0.5 * time;
+        BiFunction<Double, Double, Double> comfortFare = (distance, time) -> 1.5 * distance + 0.8 * time;
+        BiFunction<Double, Double, Double> premiumFare = (distance, time) -> 2.0 * distance + 1.5 * time;
+
+// Пример поездки
+        double distance = 10.0;  // километры
+        double time = 15.0;      // минуты
+
+// Рассчитываем стоимость для каждого тарифа
+        double economyTripCost = fareCalculator.calculateFare(distance, time, economyFare);
+        double comfortTripCost = fareCalculator.calculateFare(distance, time, comfortFare);
+        double premiumTripCost = fareCalculator.calculateFare(distance, time, premiumFare);
+
+        System.out.println("Стоимость поездки на тарифе Economy: " + economyTripCost);
+        System.out.println("Стоимость поездки на тарифе Comfort: " + comfortTripCost);
+        System.out.println("Стоимость поездки на тарифе Premium: " + premiumTripCost);
+    }
+}
